Fixed a node rendering bug.
ScriptEditorCtrl.cs
Removed the PointSize pad addition to outside the foreach loop. This fixed a bug that increased the gap between the IO sockets and bottom sockets by having more than one IO socket on the left or right (it would re-add the PointSize value to the total node hieght for each IO socket on the side-- which is why it didn't show up if there was only one IO socket on each side).

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@97 b191eb17-2330-428c-a13f-159e02ee8e09