Fix issue with nodes being difficult to select.  I added a 3 pixel tolerance to mouse movement.  this can be adjusted easily in code if it's too much or too little - but it felt good to me.
Fix VS-8 Focused field in properties panel won't update when selecting another node.  Unity won't repaint property field correctly if it still has focus.  and I can't find a way to remove focus from it - so i set focus to a different control (the search field) if the mouse is pressed down on the form

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@134 b191eb17-2330-428c-a13f-159e02ee8e09