Added special uScript_MonoDevelop.unitypackage in order to make MonoDevelop debugging work.
It seems the issue is that MonoDevelop is launching an empty Unity project and it needs all the files that are not contained in the DLL (_GUI directory, Node .cs files, and Gizmos), so I put all that into this Unity package.

INSTALL - After launching the MonoDevelop Unity debugger, import this asset package. You will only need to do this once to get the files in the right place locally.

WARNING- if any of the stuff put into the package file gets updated, this package file would also need updating (like if new nodes are added), so it is not a perfect solution.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@180 b191eb17-2330-428c-a13f-159e02ee8e09