game objects for instance sockets
ok i managed to get the changes dreamora and I have been discussing.  this is actually something I wanted to do earlier but I couldn't figure out how generate the code correctly while keeping the ui as clean as just adding a 'game object'.  anyway - it dawned on me this afternoon so I got it working.  overall it's pretty simple but it does touch all the code that deals with our event scripts and reflected entity methods so please be on the lookout for any bugs.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@411 b191eb17-2330-428c-a13f-159e02ee8e09