Made some updates to the new property grid.
The array foldout is giving me some problems, so I've commented it out for now. Added the "enabled" column that contains the checkbox to toggle whether the property value is set on the node via a socket or through the property panel.  Nothing is actually hooked up to the nodes, but the framework is there, I suppose.

I'm going to try to implement the current panel behavior using the new controls. I won't be modifying any of the existing panel behavior.  You'll need to toggle to the new panel mode still (commented out by default).

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@413 b191eb17-2330-428c-a13f-159e02ee8e09