Float Interpolation Node
Tech Details:
Added ability for nodes to have a Driven attribute on their function.  Any function with a Driven attribute is called every tick and it's outputs are pushed to the script as if it was triggered by an external connection.  These are different from Events because they don't need to be connected to GameObjects and they are transparent to the end user.

However this allows us to have interpolation nodes which push the value out each tick to any properties or variables hooked up to it.

Important Note!  This does not have all the requested features (there is no Reverse, LoopCount, LoopType or LoopDelay).  I want to make sure everything works before I go and make it more complex.  It's good enough to test with and release - when I add the new stuff it'll be upgradeable by right clicking -> upgrade node.




git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@419 b191eb17-2330-428c-a13f-159e02ee8e09