Hooked up the new uScriptGUI property controls to the actual node properties. You still have to uncomment the toggle button in DrawGUIPropertyGrid() to access the panel.
In "References.cs", added OnPaintNew() to duplicate the behavior of OnPaint() while using the new uScriptGUI controls.

When the new panel button is enabled, the new panel replaces the References panel so you can view the new and old propery grids side-by-side.
In "uScriptGUI.cs", added TextArea(), QuaternionField(), EnumTextField(), and ObjectTextField() to get one-to-one parity with the old property grid.  The only obvious issue I haven't yet addressed is the ObjectField has the height of a label even when it is associated with a RenderTexture.  It should be square in that case.  In most cases, the ObjectField should have the height of a single line of text ... but not here.

Please take a look at the new grid.  It's almost ready to go.  I haven't implemented the resizable columns yet, but it is totally usable at this point.

Almost forgot.  For now, I just force the "_tmpFieldEnabled" to TRUE.  When the field is enabled, the socket must be hidden. This doesn't currently happen.  We still need support to hide individual sockets, based on this bool checkbox.


git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@460 b191eb17-2330-428c-a13f-159e02ee8e09