Auto Packaging / Asset Referencing Added ability to auto package up the assets so they don't have to exist in the Resources folder.  To make uScript autopackage up an asset type, simply add the asset type to the uScriptConfig.cs ShouldAutoPackage function.  Right now only AudioClip is in.
For testing, or release if you guys want, I've added uScriptAct_PlaySoundDirect which takes and plays an audio clip.  It also supports drag/drop of the audio clip.

It's all functional but probably needs some sort of polish - scott and mike you guys can be the judge of that.

and again - THANKS MIKE! - for helping me past the hurdle

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@470 b191eb17-2330-428c-a13f-159e02ee8e09