Initial implementation of the Graph Contents panel.
The Node Palette panel can be toggled to display the Graph Contents. The current list is just a series of buttons under the standard categories (Events, Actions, etc). Each button is associated with a node in the graph.

Clicking a button will "select" the node. SHIFT-click will toggle the selection state of the node (which can also be used to add new nodes to the selection).

Currently, although the nodes appear to be selected in the canvas, the property grid does not display any properties for the selection.  I'm probably "selecting" the node the wrong way.

If anyone has the chance to take a look at this and point me in the right direction, please do.  Search for one of the following strings to find the location where I'm marking nodes "selected".

                                 // toggle the clicked node
                                 dn.Selected = !dn.Selected;

The panel buttons still need to be formatted. I'm going to left-align the text and color-code the buttons.  Selected buttons will use the selected node colors, for example.

Eventually, I would like to double-click a button to select the node and jump to its location on the canvas. The normal single-click should just select it (not move the canvas position).

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@594 b191eb17-2330-428c-a13f-159e02ee8e09