Fixed the ToolbarButtonStyle functionality. Updated Preference panel to use the EnumPopup instead of two checkboxes to select the style.
During the toolbar button style UpdateAll(), the code was changing the contents of a Dictionary during a foreach. It is no longer doing that. Fixed another issue that would have prevented a change to Icon only buttons.

Will style the Preference panel tomorrow or this weekend.  Thanks for implementing this!!

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@714 b191eb17-2330-428c-a13f-159e02ee8e09