Asset Package Updating - Please Read
There was a problem with our asset packaging.  It wouldn't work for assets within assets.. for example: animations within a FBX file.  So I updated it to package the asset instance as soon as the user drag/drops or selects in the property grid rather than packaging at code generation time based on the asset path.  This also means we cannot let the user manually enter a path.. so if an asset is to be packaged I turn the TextField into a LabelField.

I want everyone to be aware of how this works now just incase we encounter any weirdness, and if someone has a better idea please let me know.  PlaySound and PlayAnimation have both been updated and tested.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@745 b191eb17-2330-428c-a13f-159e02ee8e09