Initial implementation of the mini-map with zooming.
It's not pretty, but it's a start. There are some issues that need to be fixed, such as when the window is resized, but mini-map canvas size stays the same, you can click on the minimap and select nodes from the canvas, etc.

I'll fix the functionality issues first, and then address the overall look for the nodes starting with simple box colors to reflect the different nodes types. I'll probably also update the minimap background color.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@750 b191eb17-2330-428c-a13f-159e02ee8e09