Added CenterOnPoint(Point point) to support jumping to a canvas point via a mini-map click. DISCOVERED BUG - NEEDS FIX
Trapper and/or Anthony, I used CenterOnNode() as reference, and both are currently hard-coded to support a range of 0 to 4096. If you have a node with an X of more than 4096, there is no way to center on it.

I'm assuming that the 4096 was a previous hard canvas size limit. This should be updated. Please update CenterOnPoint() when CenterOnNode() is fixed.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@831 b191eb17-2330-428c-a13f-159e02ee8e09