Added a uScript information Unity editor panel that can be used for debugging. I'm currently using it to get realtime information on the window, container, and panels sizes, as well as mouse information.
It's a stand-alone editor panel that can be docked anywhere in the Unity editor interface.

In many cases, this is much easier than stepping through code to get the value of variables.  It also does not incur the performance hit from Debug.Log output, and saves the time of having to parse through the log for the info you need.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@882 b191eb17-2330-428c-a13f-159e02ee8e09