VS-300: Properties that have their associated socket exposed can no longer be edited in the property panel.  The text field has been disabled, although the default value is still visible.
ALSO: Read-only properties, such as value outputs on the Math nodes now say, "(read-only)" after the field type.

QUESTION: Should the "read only" text appear in place of the disabled text field instead? There should be no reason to show the disabled field, since it can never be edited, and will probably never reflect the actual value that will be stored in the attached variable.

Sync'd the variable names in the various Property methods so that they use the same terminology and that it's clear what the values refer to. For example, "enabled" is now "isSocketExposed".

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@996 b191eb17-2330-428c-a13f-159e02ee8e09