Initial implementation of showing deprecated nodes in the GraphContents panel.
Each deprecated node shows up PINK. For nodes that have InstanceDeprecation, there is a button that will upgrade the node when pressed.

The TypeDeprecation nodes currently have no button.

QUESTION: Is there a way to distinguish between a node that has the simple [Deprecated] attribute and one that actually points to a new class type?  My understanding is that the user should be able to "upgrade" deprecated nodes that point to a new class, but not those that just use the simple [Deprecated] attribute.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1002 b191eb17-2330-428c-a13f-159e02ee8e09