Fixed/Improved the behavior of the Preference window.
It should respond to clicks better. Some controls, such as the Save button and the new variable option would not respond to mouse clicks if the uScript editor window was too small and they overlapped other controls (under the popup window) that should have been disabled.

Added uScriptGUI.enabled to be used instead of the standard GUI.enabled. The new property checks the uScript.m_DoPreferences state to determine whether the GUI should actually be enabled.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1042 b191eb17-2330-428c-a13f-159e02ee8e09