Replaced the old context menu with the new system.  It looks a lot better.
There is an issue where after you click on an item in the new menu. The canvas goes into "canvasDragging" mode.  I've been trying to fix this for the past two hours, but I can't follow the mouse button and movement logic.

It looks like the right button (that is used to open the menu) triggers the MouseDown, but never a MouseUp. Furthermore, clicking a menu item triggers MouseDown on button 0 (left), but no MouseUp (I think), which causes it to go into the drag state.  In another section of the mouse code, it looks like the Button has a value of 4 when you click on a menu item and start dragging.  I have no idea what 4 referrs to ... maybe a combination of buttons being down??

Anyhow. It think it works pretty well.  At the moment, you just have to click a second time to get out of the drag state.  I originally was going to put in a hack so that it would avoid that state, but you would still have to click a second time to get "back on the canvas" ... which was less obvious so I commented that hack out.

Please try the new menu and see if you like it.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1055 b191eb17-2330-428c-a13f-159e02ee8e09