Updated the formatting of the PRE 3.4 drag and drop menu, moving buttons around a bit.
Only the GameObject and the actual variable type is at the base level. All other Variables associted with the dropped item are now under "Place <other>/Variables" beside any "Actions" that might exists in the menu.

Commented out a portion of DoDragDropContextMenu() that appeared to add a second "Place GameObject" that didn't actually function.

Will work on the POST 3.4 version next.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1057 b191eb17-2330-428c-a13f-159e02ee8e09