New nodes: Set Quaternion Set Random Rotation Set Random Scale Set Random Position
Note: The random rotation node has an issue where if you have a single axis preserved that happens to be at '0', you get strange math rounding errors from inside Unity's EulerAngle code I am using. If a real programmer can look at this node and find a better way to do it that would be awesome. Please note that a goal for the end user will be that they need to be able to specify in degrees (0-360) for ease in understanding.




git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1063 b191eb17-2330-428c-a13f-159e02ee8e09