Updated the ArrayFoldout<>() method to support the latest property panel layout.
Although it is called using generics, internally, it resolves to one of the "handled" object types, such as int, string, Color, and Quaternion.  The only type that isn't currently handled it Object (and GameObject?), since the way the normal uScriptGUI ObjectField is implemented looks complicated, and I didn't want to deal with it tonight.

I'ved tested with arrays of int, float, string, Vector2, and Color.  Looks cool.

I'll look at fixing the array resize issue tomorrow.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1068 b191eb17-2330-428c-a13f-159e02ee8e09