Added ResourcePathField() and AssetPathField() GUI controls to better support resource loading.
ResourcePathField() scans the entire project folder and gathers all Resources folders and their subfolders.  Presents the options as a popup list.  Using this control, the user should not be able to specify an invalid Resources location.

The AssetPathField() control provides the user with a TextField where they can type the path and asset name that they wish to reference.  A "Browse" button beside the TextField opens the AssetBrowser window.  When the user makes a selection in the window, the TextField is updated with the new asset path and name.

Implemented the AssetBrowser utilty window with the following features: You can specify the type of asset to look for. It scans the entire project for all Resources folders and then grabs the list of assets that match the type.  For example, you specify TextAsset and it generates a list of all HTML, HTM, TXT, and XML files.  The user can then choose the asset they wish to pass back to the AssetPathField().  There is some resource validation that warns the user when multiple assets share the same path and name.  The user can click on the selected asset to get it's exact path under the Assets folder, and the Project window is Ping'd.  Finally, there is a "help" icon that will open the Unity documentation in the web browser pointing to the page associated with the specific asset type.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1109 b191eb17-2330-428c-a13f-159e02ee8e09