Fixed the scrollbar spamming issue.
Wow.  Earlier, I added a custom "name" to the scrollbar styles I created.  Apparently, Unity takes that name and appends a string to it and uses it to reference scrollbar parts (like the buttons and thumb slider).   Lesson learned.  You can change the name of any style you create, but if it happens to be used on a scrollbar, watch out!

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1117 b191eb17-2330-428c-a13f-159e02ee8e09