Fix extreme lag caused when clicking the mouse (this is the performance issue I had last night)
What's happening was, everytime the mouse is clicked this is called:
string path = uScriptInstance.FindFile(uScript.Preferences.UserScripts, scriptName + ".uscript");

Which causes a very noticeable delay on our machines.

I changed it to only create the path if the button is pressed.  Mike please diff these changes and verify I did the correct thing - it all seems to function correctly when I test it.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1123 b191eb17-2330-428c-a13f-159e02ee8e09