Added some code to keep the utility window focused.
Not really keen on doing this, but at on windows, the utility window disappears behind the uScript window when the uScript window gets focus.  The only way to get to the utility window when this happens is to move the uScript window out of the way.  It's not cool.

Doesn't happen on he Mac.  There, the utility window is always on top.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1134 b191eb17-2330-428c-a13f-159e02ee8e09