Two updates:
The pre-3.4 context menu has been updated with a new styles and textures. It's position will automatically be updated to ensure that it doesn't draw outside of the uScript window. The controls in the Properties, Reference, and Scripts panels are disabled while the context menu is up to avoid that GUI issue where the click actually hits the control under the context menu first.

When you hover the mouse over Node Palette buttons, the "hot" node description appears in the Reference panel.  This was requested by a user a while ago, and something I've been meaning to do.  Users can now browse the node list and get more details on how to use them without having to actually place the node on the canvas first.

FindNodeDescription() what changed to take an EntityNode rather than a DisplayNode parameter to support the "hot" reference functionatliy.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1151 b191eb17-2330-428c-a13f-159e02ee8e09