Fixed issue where statusbar was always showing "Outside" for the mouse region.
Fixed issue where the user could drop variable nodes onto the canvas even when clicking outside of the canvas rect.  The variable would be created, but it would be hidden under a GUI panel until the canvas was panned.  Now both MouseDown and MouseUp must occur within the canvas bounds.

Fixed an issue where the user would Right-click to open the context menu and MouseDown and/or MouseUp would intermittently be triggered for the Right button when closing the menu. When this occurred, the user would have to click again to wake it up.  Now the user can immediately select and drag nodes or drag the marquee as soon as the context menu has closed.  The Right mouse button is ignored when the canvas is handling EventType.MouseDown.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1196 b191eb17-2330-428c-a13f-159e02ee8e09