The "Managed Input Event Filter" takes the specified name and grabs the current axis value out of the Input Manager array.
There are no support functions provided by Unity to test the state of that array. As far as I understand, you cannot get the array size, or test for the existance of a particular axis name.  All you can do it use GetAxis(name) to retrieve the value.  This makes it difficult or impossible to fire off event when the values change.

Fortunately, Input.anyKey is true when any mouse button, joystick button, or key is pressed, so we can pass the event to the filter.  At this time, I don't believe there is any way to fire the event by moving a joystick analog stick.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1217 b191eb17-2330-428c-a13f-159e02ee8e09