Added support for the GameObject[].  The ArrayFoldout should actually receive a string[] containing the paths/names of the GameObjects.  Internally, it deals with them as string, but knows that they are referencing GameObjects.
A custom field was created for them, with a "S" (search) button that will PING the Project Hierarchy if it finds a match. The field also supports the Drag and Drop of GameObjects onto the field.  When an object is dropped, its full path in the Hierarchy is placed in the text field.  The full paths are better, since they have a greater chance to locate the actual GO the user is looking for.

A warning message is displayed in the Console if the object they PING isn't found.  Users can test their input this way. 

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1330 b191eb17-2330-428c-a13f-159e02ee8e09