The Log Target should not show System.String[] in the "Type" column of the property grid.
There is an optional final parameter to the ArrayFoldout method that is an "override" type. Internally, the ArrayFoldout uses the Generic type specified via <>, but in the game of because I needed to be able to identify a string array as a GameObject ... this extra parameter was used.

To fix the issue, I simple removed the extra optional parameter and allowed the generic <> to handle the type.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1370 b191eb17-2330-428c-a13f-159e02ee8e09