In order to support the new-and-improved Reference panel, I need a way to access the "description" text tied to the FriendlyName attribute for node parameters.
In a previous check-in, I added uScript.FindParameterDescription(Parameter p) that should return the description tied to the passed parameter.

It looks like the Parameter data (struct) for each node is populated at launch, and it seems like the best course of action would be to add a "Description" member to ScriptEditor.Parameter struct.  I attempted to do this, but got hung up on the ParameterSerializer in ScriptEditorData.cs.  This is a bit beyond me.

Anyhow, I already added a bunch of desired Descriptions for things like EntityMethod Instance, ShowComment, and Comment parameters, as well as the Comment node parameters.  All of these lines have been commented out, but if a new "Description" member is added to the Parameter stuct, they can be uncommented and should word fine.

In short, this check-in just includes new commented statemtents that could be used in the future.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1391 b191eb17-2330-428c-a13f-159e02ee8e09