I ventured into unfamiliar territory and added the "Description" member to the Parameter stuct.
Everything seems to be working okay, but it might be a GREAT idea for Anthony and/or Trapper to check over my changes, particularly in "ScriptEditor" and "ScriptEditorData".  NOTE: I did not change the Serializer CurrentVersion from 7.  There might be some issues here.

-----

Also updated the node description text for the stuct nodes (variables, etc) and created a static class in ScriptEditor called "ParameterDescription" that just provides const strings for various parameter types, such as the descriptions for "Output Comment", "Instance", and all the comment box parameter descriptions.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1392 b191eb17-2330-428c-a13f-159e02ee8e09