Eliminated some repaints in the canvas drawing code.
Examined all the places the canvas was repainting, and identified a few areas where it appeared unnecessary.  For example, in FlowChart::RunMarqueeSelect(), all selected nodes were being deselected and then selected again.  Each node selection change called Invalidate(), which called Repaint().

There are three @FIXME comments in FlowChart.cs related to canvas, marquee, and node dragging.  Currently, it repaints once dragging begins, even if the cursor is stationary (nothing has moved).  Ideally it will repaint only when something is actually dragged or moved.  Normally uScript is ~5% CPU on my system when idle.  It jumps to ~80% while dragging.

There were a few "Selected" properties that get called numerous times with the same value.  For example, when you clicked on a node to select it.  For some reason (which might be worth looking into) Selected was being set to "false" perhaps 10 times before it was set to "true".  Each time Selected was set, methods like PrepareNode() and UpdateStyleName() were being called.  Now they are only called when Selected is toggled to the opposing value.



git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1427 b191eb17-2330-428c-a13f-159e02ee8e09