Implemented the ExportPNG.
uScriptExportPNG.cs contains all the logic. When the export is requested, a number of things happen. The panels are hidden, the map zoom is reset, and the graph bounds are determined.  Since we can only capture what is visible in the viewport, I step through the graph, one segment at a time (columns then rows).  Each segment is captured into a texture, and when all is done, the texture is encoded to PNG and then saved to disk.

There was a lot of painful tweaking to remove seams under various Unity dock layouts. I love the Unity GUI.  If you see ***ANY*** seams at all, please let me know.

The size of the graph determines the time the entire process takes.  CodeGenTest takes a number of seconds to save.  For reference, that graph generates an image that is 7,843 x 7,893 pixels in size.

I added uScriptConfig.ConstantPaths.Screenshots to point to a "Screenshots" folder in the project root.  The folder should be created one level above "Assets".  That's where the screenshots are saved.

If everyone is happy with the export functionality (process console messages, etc) ... let me know.  Should the "Export to PNG" button remain where it is, or should it be moved to the "File Menu"?  I think it should be moved to the File Menu, myself.

-----

Removed the uScript.Redraw() method, since nothing was using it.

Added uScript.RequestRepaint() method that when called, sets a bool to true. In Update(), if that bool is true, it repaints uScript.  There were numerous places where code was calling Repaint() directly, often multiple times during a single frame.  While I'm not sure Unity was actually repainting each time Repaint() was called, uScript will now repaint just once per Update(), which is what it should be doing.

I've tested the RequestRepaint() method extensively, and everything seems snappy.  If you encounter any issues, let me know.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1448 b191eb17-2330-428c-a13f-159e02ee8e09