Optimizations, App Launcing fixes and Node Parameter Descriptions This is the big checkin, I've been testing it but let me know if you find issues.
uScript.cs now has two functions:
LaunchingFromUnity - this is called when Unity launches uScript
RelaunchingFromRebuiltAppDomain - this is called when we are relaunched after Unity compiles code.
I moved all the initialization out of Update and into one of the two above (or both depending on what it was)

Changed the file list refresh to happen on Launch, AppDomain Rebuild, Save, Open.  It no longer happens every 30 seconds in the background
Added ability to find reflected node parameter description - Mike - thanks for setting all this up, sorry it took so long!

This should be a huge optimization for those with large scenes.  uScript only rebuilds it's reflected list when it's launched/relaunched instead of seemingly 'randomly' throughout using it
Also this should be huge for fixing the undo/redo problems.  Now I can tell if it was undo or just an app domain reloading.

Finally, I'm clearing the undo/redo stack (for uScript) when launching, opening a script or closing a script


git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1461 b191eb17-2330-428c-a13f-159e02ee8e09