The big Inspector checkin
You will definitely need to Remove Generated Code with this.
It's a busy weekend for me, but I talked it over with Scott and I'm going to check in.  I'll address things tonight or tomorrow night if they come up.  With a goal of putting a build out around Monday.

I tested a lot with the Alarm project and the Firebot and everything seemed good. (Besides the minor firebot change I mentioned).

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1519 b191eb17-2330-428c-a13f-159e02ee8e09