Gentlemen.  I present you with iteration 1 of real-time debugging. Steps: Make sure you save with debug info Right click nodes to add or remove breakpoints (will work on all selected nodes) Use the standard unity pause / resume controls. Breakpoints are red, current breakpoint which stopped execution are bigger and yellow
Please play with it for a couple of days and give me your feedback.
My next full uScript day is planned for Thursday and I can work on addressing big issues then.
Variable values aren't updated yet, this is next on my list.  But, for now, you can always expose them in the Inspector.

Obviously you guys will hate the breakpoint rendering, so we can work together to make it nicer.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1580 b191eb17-2330-428c-a13f-159e02ee8e09