Double-clicking on a node in the canvas will open its source file in the default editor.
NOTE: When Nested Graph nodes are double-clicked, their generated source file currently opens in the default editor.  I originally had it open the Nested Graph in the uScript Editor, replacing the current graph, but it didn't observe the dirty flag.  The user may lose work if this happens.  For now, it will just open the generated script like all other nodes.

It would be great if there were browser-line buttons on the canvas toolbar (previous and forward) so that the user could return to the previous graph after jumping into a Nested Graph node.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1584 b191eb17-2330-428c-a13f-159e02ee8e09