Nested script improvements
I really hate these checkins, it touches a lot of code and reworks some foundational stuff :-\

Logic Nodes are now technically capable of sending out parameters with their events.  I'm not making this a publicly supported feature because we don't have a UI to differentiate between parameters sent from normal output as opposed to event outputs on the same node.  However this is used by the nested scripts so events within the nested script can carry all the parameters with them and signal the parent script.

What this allows is for events in a nested script to hook up to nodes which have immediate outputs.  The code generation detects this and changes the immediate output to an event listener with all the parameters it needs.

I haven't had a chance to check all the projects, I'll be doing that over the next couple of days.  If someone runs into an error in the meantime please let me know.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1644 b191eb17-2330-428c-a13f-159e02ee8e09