RELEASE NOTES: Fixed a bug that caused multi-control properties to display incorrectly in the property grid.
DEVELOPER NOTES:
Apparently, Unity 3.5 changed EditorGUILayout.PrefixLabel() so that when you pass a empty string to it, it doesn't render (or take up space when calculating the GUI Layout.  I'm not passing " " (a space) when it would normally be empty.

This is pretty stupid.  I can see the reason why you would want to avoid unnecessary calculations, but applying an empty prefix label should still left pad the following control.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1754 b191eb17-2330-428c-a13f-159e02ee8e09