Added a VariableNameField control that allows us to know when the field loses focus.
The control duplicates the behavior of the standard TextField control, but uses reflection to access internal or sealed field, properties, and/or methods of some Unity classes.  It's not pretty, but appears to be the only way we can manually assign a ControlID and keep track of it.  I also generate a unique name for each node's control and store it in a dictionary using the controlID as the key.

During the GUI.Repaint event, I call MonitorGUIControlFocusChanges() to identify when a VariableNameField control loses focus.  It currently prints out debug text, such as "The control lost focus" ... If you want to call an event or do any special processing, that's where it should be triggered.

Note: I'm identifying a variable Name property by simply checking the friendly name.  If it matches "Name", VariableNameControl is called instead of TextField.  This might not be the best way to go, since non-variable nodes might also have friendly names of "Name".  If there is a better way to identify the Name property of a variable node, please update this.

-----

A few non-related changes include moving some utility methods to the uScriptUtility class, and replacing simple properties in with c# auto-properties to simplify the code.


git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@2096 b191eb17-2330-428c-a13f-159e02ee8e09