Added a keyboard shortcut (ALT-G) to toggle grid snapping.  The shortcut to toggle the grid visibility (CTRL-G) remains the same.
Updated the canvas toolbar, adding a Grid Snap toggle button, and buttons for the individual save modes.

Updated the keyboard shortcut logic so that the variable shortcut keys only work when no modifier buttons are held.  It was the case that you could hold down CTRL-ALT-B (or any other modifier combo) and click to place a Bool variable.  Now, you must hold B only (with no modifier).  This frees up the modifiers to be used for other shortcuts in the future.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@2157 b191eb17-2330-428c-a13f-159e02ee8e09