Initial pass on Toolbox Favorites
Select a node on the canvas, and a Favorites badge will appear at the upper-right corner of the Reference panel.  It defaults to "-".  Click the badge and then select a number (1-9) from the popup to assign it a favorite keyboard shortcut.  The user will now be able to place this node by holding down the specified number key and clicking on the canvas, just like the common variables (B for Boolean).

When one or more favorites have been assigned, a small section appears at the bottom of the Toolbox panel.  This section shows the current favorites, including the assigned number.  Clicking the node button will place the node on the canvas, just like clicking any other node button in the Toolbox.

To remove a favorite shortcut assignment, you can either select a node of that type on the canvas, and simply change the shortcut badge back to "-", or you can click the "Edit" button on the "Favorite Nodes" toolbar.  That will enter "edit mode".  Clicking a favorite node button while in edit mode will remove the shortcut.  This is the easiest way to clear the favorite assignments.

It's all pretty functional, but not pretty in some areas.  Tomorrow I will be doing the following:
+ Changing the color of the buttons while in edit mode (they are currently too red)
+ Fixing the button widths, so they look correct in all cases (currently, they match the Comment node width)
+ Moving the Favorite Badge from the Reference panel, and placing it on the node foldout button bar in the Property panel (this seems like a more logical place to find it)

All the common node types are already supported (actions, events, variables, comments, etc), and I'll be adding support for the rest tomorrow (OwnerConnection and reflected methods and properties).

I think that's about it.  If you get the chance to play with it, let me know what you think.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@2181 b191eb17-2330-428c-a13f-159e02ee8e09