Added a few improvements.  Will send to Unity support for review.
The time it takes for the menu to appear under either OS is relative to the number of items the menu contains.
On Mac, a menu containing 8,000 items appears after a delay of approximately 1 second.
On Windows, the same menu appears after a delay of approximately 24 seconds.

For reference, our context menu contains fewer than 1,500 items.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@2314 b191eb17-2330-428c-a13f-159e02ee8e09