Fixed some instances of a renamed variable that didn't get updated.
Mike- I assume this was you using one of your fancy Visual Studio add-ons. As an FYI-- it doesn't seem to update names for things that are IfDef'ed out. I've seen this in the past and meant to tell you. The same is true when you rename actual .cs files-- any code calling those classes that are IfDef'ed out do NOt get updated as part of you add-ons rename proccess.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@2424 b191eb17-2330-428c-a13f-159e02ee8e09